+++ completions/npm/hooks.ps1
@@ -8,7 +8,7 @@ function handleCompletions($completions) {
     $devDependencies = $packageJson.devDependencies
 
     $input_arr = $PSCompletions.input_arr
-    $filter_input_arr = $PSCompletions.filter_input_arr # Exclude options parameters
+    $filter_input_arr = $PSCompletions.filter_input_arr # Exclude option parameters
     $first_item = $filter_input_arr[0] # The first subcommand
     $last_item = $filter_input_arr[-1] # The last subcommand