+++ module/PSCompletions/PSCompletions.psm1
@@ -101,7 +101,7 @@ function PSCompletions {
     }
     function Out-Data {
         if ($PSCompletions.__need_update_data) {
-            $PSCompletions.data | ConvertTo-Json -Compress -Depth 10 | Out-File $PSCompletions.path.data -Force -Encoding utf8
+            $PSCompletions.data | ConvertTo-Json -Depth 10 | Out-File $PSCompletions.path.data -Force -Encoding utf8
             $PSCompletions.__need_update_data = $null
         }
     }
@@ -229,7 +229,7 @@ function PSCompletions {
             $data.config.comp_config.Remove($_)
         }
 
-        $data | ConvertTo-Json -Compress -Depth 10 | Out-File $PScompletions.path.data -Force -Encoding utf8
+        $data | ConvertTo-Json -Depth 10 | Out-File $PScompletions.path.data -Force -Encoding utf8
         $PSCompletions.data = $data
     }
     function _update {
@@ -513,7 +513,7 @@ function PSCompletions {
         }
         $PSCompletions.data.alias = $data_alias
         $PSCompletions.data.aliasMap = $data_aliasMap
-        $PSCompletions.data | ConvertTo-Json -Compress -Depth 10 | Out-File $PScompletions.path.data -Force -Encoding utf8
+        $PSCompletions.data | ConvertTo-Json -Depth 10 | Out-File $PScompletions.path.data -Force -Encoding utf8
         if ($need_restart) {
             $PSCompletions.write_with_color((_replace $PSCompletions.info.module.restart))
             $need_restart = $null