+++ module/PSCompletions/core.ps1
@@ -8,7 +8,7 @@ if ($PSCompletions.guid) {
 
 $_ = $PSScriptRoot
 New-Variable -Name PSCompletions -Option Constant -Value @{
-    version                 = '6.6.4'
+    version                 = '6.7.0'
     path                    = @{
         root             = $_
         completions      = "$_\completions"